Authentication on restricted actions

+++ src/main/java/com/qa/choonz/service/AlbumService.java
@@ -1,32 +1,33 @@
 package com.qa.choonz.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import com.qa.choonz.utils.ActiveSessions;
-import org.springframework.stereotype.Service;
-
 import com.qa.choonz.exception.AlbumNotFoundException;
 import com.qa.choonz.persistence.domain.Album;
+import com.qa.choonz.persistence.domain.UserDetails;
 import com.qa.choonz.persistence.repository.AlbumRepository;
+import com.qa.choonz.persistence.roles.UserRole;
 import com.qa.choonz.rest.dto.AlbumDTO;
 import com.qa.choonz.rest.mapper.AlbumMapper;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class AlbumService {
 
     private AlbumRepository repo;
     private AlbumMapper mapper;
-    private ActiveSessions sessions;
 
-    public AlbumService(AlbumRepository repo, AlbumMapper mapper, ActiveSessions sessions) {
+    public AlbumService(AlbumRepository repo, AlbumMapper mapper) {
         super();
         this.repo = repo;
         this.mapper = mapper;
-        this.sessions = sessions;
     }
 
-    public AlbumDTO create(Album album) {
+    public AlbumDTO create(Album album, UserDetails user) {
+        // ensure only admin can create albums
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Album created = this.repo.save(album);
         return mapper.mapToDeepDTO(created);
     }
@@ -40,11 +41,10 @@ public class AlbumService {
         return mapper.mapToDeepDTO(found);
     }
 
-    public List<AlbumDTO> readByAlbum(long id) {
-        return this.repo.findAllByArtist_Id(id).stream().map(mapper::mapToDeepDTO).collect(Collectors.toList());
-    }
+    public AlbumDTO update(Album album, long id, UserDetails user) {
+        // ensure only admin can edit albums
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
 
-    public AlbumDTO update(Album album, long id) {
         Album toUpdate = this.repo.findById(id).orElseThrow(AlbumNotFoundException::new);
         toUpdate.setName(album.getName());
         toUpdate.setArtist(album.getArtist());
@@ -53,7 +53,10 @@ public class AlbumService {
         return mapper.mapToDeepDTO(updated);
     }
 
-    public boolean delete(long id) {
+    public boolean delete(long id, UserDetails user) {
+        // ensure only admin can delete albums
+        if (user == null || user.getRole() != UserRole.ADMIN){return false;}
+
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
